SceneSpaceType changed back to float again
git-svn-id: https://svn.vis.uni-stuttgart.de/utilities/vislib/trunk@390 3ec011a9-ccf5-4689-8f2f-5f4aa021d377

+++ graphics/include/vislib/graphicstypes.h
@@ -18,7 +18,7 @@ namespace vislib {
 namespace graphics {
 
     /** This type is used for values in scene space. */
-    typedef double SceneSpaceType;
+    typedef float SceneSpaceType;
 
     /** 
      * This type is used for values in image space.
